build your army menu
add text to menu and mad it so when buttons click shows how many of each unit type would be bought. still need to add a abckground to the menus to make the easier to read. as well maki it so when you buy units it can be used for batteling